Persistence-related fixes and improvements
Add a configuration option to use message persistence. The default is to
not use persistence; if set to "true" in-memory persistence will be
used.
Fix the client using the default file persistence for admin messages,
resulting in errors in deployments where temporary files could not be
opened - typically in containers.
Document deployment steps for quarkus-based versions of Keycloak.

Change-Id: I8430b4b9c8aad536a6f53090a1a14aea7fad5546